new_edit

+++ app/Http/Controllers/IdeaController.php
@@ -10,7 +10,8 @@ class IdeaController extends Controller
 {
     public function store()
     {
-        request()->validate(
+
+       $validated = request()->validate(
             [
                 'content' => 'required'
             ],
@@ -18,13 +19,11 @@ class IdeaController extends Controller
                 'content.required'=>'This is required'
             ]
         );
-        $idea =  Idea::create(
-            [
-                'content' => request()->get('idea', ''),
-            ],
-        );
+        $validated['user_id']=auth()->id();
 
-        return redirect()->route('dashboard')->with('success', 'ok');
+         Idea::create($validated);
+
+        return redirect()->route('dashboard')->with('success', 'Shared successfully');
     }
 
     public function show(Idea $idea)
@@ -34,11 +33,18 @@ class IdeaController extends Controller
 
     public function edit(Idea $idea)
     {
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
         $editing = true;
         return view('shared.show-idea', compact('idea','editing'));
     }
 
     public function update(Idea $idea){
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
         request()->validate(
             [
                 'content' => 'required'
@@ -62,9 +68,14 @@ class IdeaController extends Controller
 
     public function destroy(Idea $idea)
     {
+
+        if(auth()->id() !== $idea->user_id){
+            abort(404);
+        }
+
         try {
             $idea->delete();
-            return redirect()->back()->with('delete', 'Deleted successfully');
+            return redirect()->back()->with('success', 'Deleted successfully');
         } catch (Exception $exception) {
             return abort(404);
         }
